// projects/bible/mobile/src/interfaces/ISongsBook.ts
// Generated by https://quicktype.io

export interface ISongsBookResponse {
  results: number;
  code: number;
  status: string;
  ok: boolean;
  data: ISong[];
}

export interface ISong {
  id: number;
  uuid: string;
  views: number;
  likes: number;
  title: string;
  num: number;
  songId: string;
  lyrics: string;
  lyricsHtml: string;
  slug: string;
  video: null;
  language: Language;
  SongCategoryId: number;
  createdAt: string;
  updatedAt: string;
  SongCategory: SongCategory;
}

export interface SongCategory {
  id: number;
  uuid: string;
  name: string;
  slug: string;
  abbreviation: string;
  createdAt: string;
  updatedAt: string;
}

export enum Language {
  Fr = 'fr',
  HT = 'ht',
}